+++ locales/en/search.json
@@ -5,7 +5,6 @@
     "language-select": "Select a language",
     "reset": "reset",
     "search": "Search",
-    "search-results": "{{ count }} search results",
     "search-desc": "Search the corpus of the Holy Quran and its translations using a powerful and highly accurate multi-lingual search engine.",
     "search-title": "Search for '{{searchQuery}}'",
     "translation-select": "Select a translation"
