--- src/components/Search/NavigationItem/NavigationItem.module.scss
@@ -1,27 +1,0 @@
-.container {
-  display: inline-flex;
-  align-items: center;
-  border-radius: var(--border-radius-default);
-}
-
-.name {
-  font-size: var(--font-size-normal);
-  line-height: var(--line-height-large);
-  padding-block-start: var(--spacing-micro);
-  padding-block-end: var(--spacing-micro);
-  padding-inline-start: var(--spacing-micro);
-  padding-inline-end: var(--spacing-micro);
-}
-
-.link {
-  text-decoration: none;
-}
-
-.key {
-  border-radius: var(--border-radius-default);
-  border: 1px solid var(--color-background-alternative-deep);
-  padding-block-start: var(--spacing-micro);
-  padding-block-end: var(--spacing-micro);
-  padding-inline-start: var(--spacing-micro);
-  padding-inline-end: var(--spacing-micro);
-}
