+++ src/components/dls/Popover/index.tsx
@@ -34,7 +34,6 @@ interface Props {
   triggerStyles?: string;
   contentStyles?: string;
   contentSideOffset?: number;
-  className?: string;
 }
 
 const Popover: React.FC<Props> = ({
@@ -53,9 +52,8 @@ const Popover: React.FC<Props> = ({
   contentSideOffset = 2,
   triggerStyles,
   contentStyles,
-  className,
 }) => (
-  <div className={classNames({ [styles.container]: defaultStyling }, className)}>
+  <div className={classNames({ [styles.container]: defaultStyling })}>
     <RadixPopover.Root
       modal={isModal}
       {...(typeof open !== 'undefined' && { open })}
